+++ CHANGELOG.md
@@ -18,6 +18,7 @@ All notable changes to this project will be documented in this file. The format 
 ### Changed
 - Allow field containers to be included in the list of `fields` in `FieldContainer(fields)`. If a field container is included, its sub-fields are unpacked.
 - Set custom attributes by keyword-arguments in `FieldContainer(fields, **kwargs)`.
+- Show also the progress of Newton's method for `Job.evaluate(verbose=True)`.
 
 ## [9.3.0] - 2025-06-29